Created value and result struct
