+++ lib/arel/middleware/postgresql.rb
@@ -1,5 +1,7 @@
 require_relative './postgresql/ffi'
 require_relative './postgresql/ffi/column'
+require_relative './postgresql/ffi/value'
+require_relative './postgresql/ffi/result'
 require_relative './postgresql/adapter'
 require_relative './postgresql/result'
